MADE DESIGN BEAUTIFUL
I mostly kept the changes I made in the css to make the web page look ok
and modified the state in the MapContainer component.

Some issues :
1. I've proved unable to fix this first one. The marker appears on the
map but its icon won't appear. I haven't yet understood why. The popup
does show up too and you get the right output when you click on it.

2. Warning: Each child in a list should have a unique "key" prop.
-> I know it's not a big deal as it won't prevent react from working
properly but I haven't been able to fix it yet too. This isn't a
priority.

3. I need to display json data on my map. I managed to get the json data
in the project but now I think I should loop through it to get the
various restaurants marked on the map.